requirements.txt
Django==3.1
psycopg2-binary==2.9.1
Pillow==8.3.1
mathfilters
django-crispy-forms
django-shopping-cart